Split apart updateMatchState into updateMatchStateNarrow and updateMatchStateWide.
This speeds up the code slightly due to the need to do fewer
evaluations, but more importantly it opens the door to using different
shortcuts in a wide and narrow context. This is helpful because extended
Latin is full of ambiguous characters, making it impossible to write
efficient shortcuts for languages using diacritics in Latin script.