Improve Attributed string performance
Replaces the naïve implementation of `Attributed a` as a cons-based data
structure with a `Seq` of inner `Attr` elements. This eliminates the
excessive recursion when calling `attrRender` and `attrStrip` to render
the attributed string output to the printable output in my previous
iteration.

Benchmarks
==========

Timings from an M2 Max Mac Studio for the "reflow" and "tabular"
benchmarks here and some representative writer benchmarks from Pandoc.
I added the contents of `tables.markdown` to `testsuite.txt` for these
runs in hopes that it would give a little extra exercise to `cblock`
etc.

(I didn't run the pandoc benchmarks with the previous slow
implementation.)

The `Seq` column is timings using this commit. The list column is from a
version where the `Attributed a` is a `[Attr a]` rather than a `Seq
(Attr a)`, which I compared just to smell-test my choice of data
structure. The list version performed slightly better than the `Seq`
version on the tabular benchmarks, which heavily exercise DocLayout's
`Block` support, which employs prerendering. However it performs worse
on the reflow benchmarks and on the pandoc writers, where the present
commit actually shows a slight performance improvement over baseline.

None of the pandoc writers seem to employ extravagant uses of `lblock`,
`vfill`, and friends. It's possible that nested tables (are those even
supported?) in a source document could hit the slow behavior shown by
the tabular benchmarks, but I think they might basically be unrealistic
as to common interactive Pandoc usage.

I don't really know whether using `[Attr a]` or `Seq (Attr a)` is
ultimately "better" for our purposes. As long as the thing you get when
you unwrap `Attributed` is a `Foldable` of `Attr` there should be no
further need to fiddle with the rendering functions if anyone wanted to
experiment more with other implementations.

| Benchmark           | Before ANSI | cons     | `Seq`    | list     |
|---------------------|-------------|----------|----------|----------|
| reflow English      | 110 µs      | 244 µs   | 109 µs   | 142 µs   |
| reflow Greek        | 98 µs       | 321 µs   | 100 µs   | 126 µs   |
| tabular English     | 1.347 ms    | 4.751 ms | 2.623 ms | 1.911 ms |
| tabular Greek       | 1.711 ms    | 6.462 ms | 2.893 ms | 2.283 ms |
| asciidoctor writer  | 2.83 ms     | -        | 2.77 ms  | 2.85 ms  |
| docbook5 writer     | 2.91 ms     | -        | 2.75 ms  | 2.93 ms  |
| opendocument writer | 7.00 ms     | -        | 6.71 ms  | 7.08 ms  |