Introduce ANSI link support
The `Attr` constructor grows a field for a link target, which is just a
`Maybe Text` for the time being. We deem nested links to be illegal,
like HTML, and so the rendering state only has to keep track of one link
at a time. When rendering to ANSI, we use OSC8 to include the link in
the output.

OSC8 links support an "id" parameter which tells terminals to treat
discontiguous linked text spans as being "the same" link, for mouseover
purposes. Our code could grow to support this.