Use Data.Text.Lazy.Builder in ANSI rendering
We have to smush a bunch of strings together in a fold with auxiliary
data and hit a quadratic time blowup. This change means we return a lazy
Text from renderANSI instead of whatever the contents of `Doc a` were
but in the relevant contexts I don't think anyone will mind.