Use Data.Text.Lazy.Builder in ANSI rendering (#29)
We have to smush a bunch of strings together in a fold with auxiliary
data and hit a quadratic time blowup. This change means we return a lazy
Text from renderANSI instead of whatever the contents of `Doc a` were
but in the relevant contexts I don't think anyone will mind.

API change: added a `build` method to the signature of HasChars.
This has a default definition, and is only needed with the ANSI
renderer, so existing users should not need to add anything to
their HasChars instances.